# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Protomind API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/meetings": {
            "get": {
                "tags": [
                    "Meetings"
                ],
                "summary": "Показать список совещаний",
                "description": "Метод для отображения списка совещаний",
                "operationId": "indexMeetings",
                "responses": {
                    "200": {
                        "description": "Успешная операция",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MeetingResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "Meetings"
                ],
                "summary": "Создать новое совещание",
                "description": "Метод для создания нового совещания",
                "operationId": "storeMeeting",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MeetingStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешная операция",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MeetingResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибочный запрос",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ошибка при создании совещания"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/meetings/{id}": {
            "get": {
                "tags": [
                    "Meetings"
                ],
                "summary": "Показать совещание",
                "description": "Метод для отображения данных совещания",
                "operationId": "showMeeting",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID совещания",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешная операция",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MeetingResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Совещание не найдено."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Meetings"
                ],
                "summary": "Обновить совещание",
                "description": "Метод для обновления совещания",
                "operationId": "updateMeeting",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID совещания",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MeetingUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешная операция",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MeetingResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Совещание не найдено."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Meetings"
                ],
                "summary": "Удалить совещание",
                "description": "Метод для удаления совещания",
                "operationId": "destroyMeeting",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID совещания",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Совещание не найдено."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MeetingStoreRequest": {
                "title": "Запрос на создание совещания",
                "description": "Запрос, содержащий данные для создания совещания",
                "properties": {
                    "secretary_id": {
                        "description": "ID секретаря",
                        "type": "integer"
                    },
                    "document": {
                        "description": "Документ совещания",
                        "type": "string",
                        "format": "binary"
                    },
                    "theme": {
                        "description": "Тема совещания",
                        "type": "string",
                        "maxLength": 255
                    },
                    "link": {
                        "description": "Ссылка на совещание",
                        "type": "string",
                        "maxLength": 255
                    },
                    "event_date": {
                        "description": "Дата совещания",
                        "type": "string",
                        "format": "date"
                    },
                    "event_start_time": {
                        "description": "Время начала совещания",
                        "type": "string",
                        "format": "date-time"
                    },
                    "event_end_time": {
                        "description": "Время окончания совещания",
                        "type": "string",
                        "format": "date-time"
                    },
                    "members": {
                        "description": "Список участников совещания",
                        "type": "array",
                        "items": {
                            "properties": {
                                "member_id": {
                                    "description": "ID участника",
                                    "type": "integer"
                                },
                                "should_notify": {
                                    "description": "Уведомлять ли участника",
                                    "type": "boolean"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "MeetingUpdateRequest": {
                "title": "Запрос на обновление совещания",
                "description": "Запрос, содержащий данные для обновления совещания",
                "properties": {
                    "theme": {
                        "description": "Тема совещания",
                        "type": "string",
                        "maxLength": 255,
                        "nullable": true
                    },
                    "document": {
                        "description": "Документ совещания",
                        "type": "string",
                        "format": "binary",
                        "nullable": true
                    },
                    "link": {
                        "description": "Ссылка на совещание",
                        "type": "string",
                        "maxLength": 255,
                        "nullable": true
                    },
                    "event_date": {
                        "description": "Дата совещания",
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "event_start_time": {
                        "description": "Время начала совещания",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "event_end_time": {
                        "description": "Время окончания совещания",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "members": {
                        "description": "Список участников совещания",
                        "type": "array",
                        "items": {
                            "properties": {
                                "member_id": {
                                    "description": "ID участника",
                                    "type": "integer"
                                },
                                "should_notify": {
                                    "description": "Уведомлять ли участника",
                                    "type": "boolean"
                                }
                            },
                            "type": "object"
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "MeetingMemberResource": {
                "title": "Ресурс участника совещания",
                "description": "Ресурс, представляющий участника совещания",
                "properties": {
                    "id": {
                        "description": "ID участника совещания",
                        "type": "integer"
                    },
                    "member": {
                        "$ref": "#/components/schemas/UserResource"
                    },
                    "email_sent": {
                        "description": "Флаг отправки email уведомления",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "MeetingResource": {
                "title": "Ресурс совещания",
                "description": "Ресурс, представляющий совещание",
                "properties": {
                    "id": {
                        "description": "ID совещания",
                        "type": "integer"
                    },
                    "theme": {
                        "description": "Тема совещания",
                        "type": "string"
                    },
                    "link": {
                        "description": "Ссылка на совещание",
                        "type": "string"
                    },
                    "event_date": {
                        "description": "Дата совещания",
                        "type": "string",
                        "format": "date"
                    },
                    "event_start_time": {
                        "description": "Время начала совещания",
                        "type": "string",
                        "format": "date-time"
                    },
                    "event_end_time": {
                        "description": "Время окончания совещания",
                        "type": "string",
                        "format": "date-time"
                    },
                    "members": {
                        "description": "Список участников совещания",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/MeetingMemberResource"
                        }
                    },
                    "document_path": {
                        "description": "Путь к документу совещания",
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "UserResource": {
                "title": "Пользователь",
                "description": "Ресурс пользователя",
                "properties": {
                    "id": {
                        "description": "ID пользователя",
                        "type": "integer"
                    },
                    "full_name": {
                        "description": "Полное имя пользователя",
                        "type": "string"
                    },
                    "email": {
                        "description": "Электронная почта пользователя",
                        "type": "string",
                        "format": "email"
                    },
                    "login": {
                        "description": "Логин пользователя",
                        "type": "string"
                    },
                    "role": {
                        "description": "Роль пользователя",
                        "type": "string"
                    },
                    "department": {
                        "description": "Отдел пользователя",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Дата и время создания пользователя",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Дата и время последнего обновления пользователя",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
...
    }
}